+++ .github/workflows/pythonpackage.yml
@@ -100,8 +100,8 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-#        pip install numpy scipy h5py cython pandas xlrd pytest ipywidgets IPython matplotlib traitlets pyqt5
-#        pip install uncertainties ptemcee corner tqdm pytest-qt periodictable wheel delocate pyopencl
+        pip install numpy scipy h5py cython pandas xlrd pytest ipywidgets IPython matplotlib traitlets pyqt5
+        pip install uncertainties ptemcee corner tqdm pytest-qt periodictable wheel delocate pyopencl
         pip install git+https://github.com/pymc-devs/pymc3
 
     - name: Make wheel
@@ -116,8 +116,10 @@ jobs:
       env:
         MPLBACKEND: agg
       run: |
-        pip install --only-binary=refnx --no-index --find-links=dist/ refnx
-        tox
+        pushd dist
+        pip install --only-binary=refnx --no-index refnx
+        python -c "import refnx;refnx.test()"
+        popd
 
     - name: Make frozen GUI executable
       if: ${{ matrix.python-version == '3.8' }}
