+++ MANIFEST.in
@@ -1,4 +1,3 @@
 # All source files
 recursive-include refnx *
 recursive-include src *
-include _open_mp_helpers.py
